Use size_t for return type to size() rather than int

+++ org/antlr/v4/runtime/ANTLRInputStream.h
@@ -115,7 +115,7 @@ namespace org {
                     /// </summary>
                     virtual int index() override;
 
-                    virtual int size() override;
+					virtual size_t size() override;
 
                     /// <summary>
                     /// mark/release do nothing; we have entire buffer </summary>
